@@ -22,7 +22,7 @@
     "Veselin Todorov <hi@vesln.com>",
     "John Firebaugh <john.firebaugh@gmail.com>"
   ],
-  "version": "6.2.0",
+  "version": "6.2.1",
   "repository": {
     "type": "git",
     "url": "https://github.com/chaijs/chai"
@@ -32,7 +32,7 @@
   },
   "main": "./index.js",
   "scripts": {
-    "build": "esbuild --bundle --format=esm --keep-names --outfile=index.js lib/chai.js",
+    "build": "esbuild --bundle --format=esm --target=es2021 --keep-names --outfile=index.js lib/chai.js",
     "prebuild": "npm run clean",
     "format": "prettier --write lib",
     "pretest": "npm run lint",
@@ -50,22 +50,22 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
-    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-commonjs": "^29.0.0",
     "@web/dev-server-rollup": "^0.6.1",
-    "@web/test-runner": "^0.18.0",
+    "@web/test-runner": "^0.20.0",
     "@web/test-runner-playwright": "^0.11.0",
     "assertion-error": "^2.0.1",
     "c8": "^10.1.3",
     "check-error": "^2.1.1",
     "deep-eql": "^5.0.1",
-    "esbuild": "^0.25.9",
-    "eslint": "^8.56.0",
-    "eslint-plugin-jsdoc": "^48.0.4",
+    "esbuild": "^0.27.0",
+    "eslint": "^9.0.0",
+    "eslint-plugin-jsdoc": "^61.0.0",
     "globals": "^16.3.0",
     "loupe": "^3.1.0",
-    "mocha": "^10.2.0",
+    "mocha": "^11.0.0",
     "pathval": "^2.0.0",
     "prettier": "^3.4.2",
-    "typescript": "~5.7.3"
+    "typescript": "~5.9.0"
   }
 }
