@@ -178,34 +180,36 @@
     "date-time": "^4.0.0",
     "es5-shim": "^4.6.7",
     "es6-shim": "^0.35.8",
-    "eslint": "^9.37.0",
+    "eslint": "^9.39.1",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.4",
-    "eslint-plugin-unicorn": "^61.0.2",
-    "eslint-plugin-vue": "^10.5.0",
+    "eslint-plugin-unicorn": "^62.0.0",
+    "eslint-plugin-vue": "^10.5.1",
     "fixturify": "^3.0.0",
     "flru": "^1.0.2",
     "fs-extra": "^11.3.2",
     "github-api": "^3.4.0",
-    "globals": "^16.4.0",
+    "globals": "^16.5.0",
     "husky": "^9.1.7",
     "is-reference": "^3.0.3",
-    "lint-staged": "^16.2.4",
+    "lint-staged": "^16.2.6",
     "locate-character": "^3.0.0",
-    "magic-string": "^0.30.19",
-    "memfs": "^4.49.0",
-    "mocha": "^11.7.4",
-    "nodemon": "^3.1.10",
+    "magic-string": "^0.30.21",
+    "memfs": "^4.50.0",
+    "mocha": "^11.7.5",
+    "nodemon": "^3.1.11",
+    "npm-audit-resolver": "^3.0.0-RC.0",
     "nyc": "^17.1.0",
+    "patch-package": "^8.0.1",
     "picocolors": "^1.1.1",
     "picomatch": "^4.0.3",
-    "pinia": "^3.0.3",
+    "pinia": "^3.0.4",
     "prettier": "^3.6.2",
     "prettier-plugin-organize-imports": "^4.3.0",
     "pretty-bytes": "^7.1.0",
     "pretty-ms": "^9.3.0",
     "requirejs": "^2.3.7",
-    "rollup": "^4.52.4",
+    "rollup": "^4.53.2",
     "rollup-plugin-license": "^3.6.0",
     "rollup-plugin-string": "^3.0.0",
     "semver": "^7.7.3",
